pyproject.toml
[project]
name = "mcp-github-app"
version = "0.1.0"
description = "MCP server (Python) for GitHub using GitHub App auth"
requires-python = ">=3.10"
dependencies = [
  "modelcontextprotocol>=0.1.0",
  "pyjwt[crypto]",
  "cryptography",
  "requests",
  "GitPython",
  "pydantic",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mcp_github_app"]